refactor(visualization): open a row's context menu through one service
Opening the node context menu from a row — the guard, the event, the
scroll listener that closes it again — moves out of the file tree's row
component into an ExplorerRowContextMenuService, so any row an explorer
lists can offer the menu without repeating it.
