+++ visualization/app/codeCharta/features/sidebarExplorer/components/explorerTreeLevel/explorerTreeLevel.component.ts
@@ -1,14 +1,13 @@
-import { ChangeDetectionStrategy, Component, computed, DestroyRef, effect, inject, input, OnInit, signal } from "@angular/core"
+import { ChangeDetectionStrategy, Component, computed, effect, inject, input, OnInit, signal } from "@angular/core"
 import { CodeMapNode } from "../../../../model/codeCharta.model"
 import { isLeaf } from "../../../../util/codeMapHelper"
-import { EXPLORER_CONTEXT_MENU } from "../../explorerContextMenu"
 import { EXPLORER_ROW } from "../../explorerRow"
 import { explorerRowId } from "../../explorerRowId"
 import { EXPLORER_SELECTION } from "../../explorerSelection"
 import { EXPLORER_STORAGE_SCOPE } from "../../explorerStorageScope"
 import { scrollRowIntoViewWhenRendered } from "../../scrollRowIntoView"
 import { ExplorerRevealService } from "../../services/explorerReveal.service"
-import { ExplorerScrollHostService } from "../../services/explorerScrollHost.service"
+import { ExplorerRowContextMenuService } from "../../services/explorerRowContextMenu.service"
 import { ExplorerRowComponent } from "../explorerRow/explorerRow.component"
 import { ExplorerTreeItemIconComponent } from "../explorerTreeItemIcon/explorerTreeItemIcon.component"
 import { ExplorerTreeItemNameComponent } from "../explorerTreeItemName/explorerTreeItemName.component"
@@ -22,13 +21,9 @@ import { ExplorerTreeItemNameComponent } from "../explorerTreeItemName/explorerT
 export class ExplorerTreeLevelComponent implements OnInit {
     private readonly row = inject(EXPLORER_ROW)
     private readonly selection = inject(EXPLORER_SELECTION)
-    private readonly contextMenu = inject(EXPLORER_CONTEXT_MENU, { optional: true })
+    private readonly contextMenu = inject(ExplorerRowContextMenuService)
     private readonly revealService = inject(ExplorerRevealService)
-    private readonly scrollHostService = inject(ExplorerScrollHostService)
     private readonly storageScope = inject(EXPLORER_STORAGE_SCOPE)
-    private readonly destroyRef = inject(DestroyRef)
-
-    private scrollHostWithListener: HTMLElement | null = null
 
     readonly node = input.required<CodeMapNode>()
     readonly depth = input.required<number>()
@@ -40,8 +35,7 @@ export class ExplorerTreeLevelComponent implements OnInit {
     readonly isSelectable = computed(() => this.rowProjection().isSelectable)
     readonly isHovered = computed(() => this.selection.isHovered(this.node()))
     readonly isSelected = computed(() => this.selection.isSelected(this.node()))
-    readonly isMarked = computed(() => this.contextMenu?.isMarked(this.node().path) ?? false)
-    readonly hasContextMenu = computed(() => this.contextMenu?.isEnabledFor(this.node().path) ?? false)
+    readonly isMarked = computed(() => this.contextMenu.isMarked(this.node().path))
     readonly isRevealed = computed(() => this.revealService.revealedNodePath() === this.node().path)
     readonly isLeafNode = computed(() => isLeaf(this.node()))
     readonly children = computed(() => this.node().children ?? [])
@@ -64,7 +58,6 @@ export class ExplorerTreeLevelComponent implements OnInit {
 
     ngOnInit(): void {
         this.isOpen.set(this.depth() === 0)
-        this.destroyRef.onDestroy(() => this.removeScrollListener())
     }
 
     onMouseEnter($event: MouseEvent) {
@@ -93,34 +86,6 @@ export class ExplorerTreeLevelComponent implements OnInit {
     }
 
     openNodeContextMenu = ($event: MouseEvent) => {
-        if (!this.hasContextMenu()) {
-            return
-        }
-
-        $event.preventDefault()
-        $event.stopPropagation()
-
-        this.contextMenu?.open(this.node().path, $event.clientX, $event.clientY)
-
-        this.addScrollListener()
-    }
-
-    private addScrollListener() {
-        if (this.scrollHostWithListener) {
-            return
-        }
-        const scrollHost = this.scrollHostService.element()
-        scrollHost?.addEventListener("scroll", this.closeContextMenuOnScroll)
-        this.scrollHostWithListener = scrollHost
-    }
-
-    private removeScrollListener() {
-        this.scrollHostWithListener?.removeEventListener("scroll", this.closeContextMenuOnScroll)
-        this.scrollHostWithListener = null
-    }
-
-    private readonly closeContextMenuOnScroll = () => {
-        this.contextMenu?.close()
-        this.removeScrollListener()
+        this.contextMenu.openFor(this.node().path, $event)
     }
 }
